+++ angular-workspace/projects/ni/nimble-angular/src/directives/anchor/nimble-anchor.directive.ts
@@ -1,8 +1,8 @@
 import { Directive, ElementRef, Input, Renderer2 } from '@angular/core';
 import type { Anchor } from '@ni/nimble-components/dist/esm/anchor';
 import { AnchorAppearance } from '@ni/nimble-components/dist/esm/anchor/types';
+import { BooleanValueOrAttribute, toBooleanProperty } from '@ni/nimble-angular/internal-utilities';
 import { NimbleAnchorBaseDirective } from '../anchor-base/nimble-anchor-base.directive';
-import { BooleanValueOrAttribute, toBooleanProperty } from '../utilities/template-value-helpers';
 
 export type { Anchor };
 export { AnchorAppearance };