+++ packages/nimble-components/src/patterns/dropdown/styles.ts
@@ -130,19 +130,19 @@ export const styles = css`
         border-bottom-color: ${failColor};
     }
 
+    .anchored-region[hidden] {
+        visibility: hidden;
+    }
+
     .listbox {
         box-sizing: border-box;
         display: inline-flex;
         flex-direction: column;
         left: 0;
         overflow-y: auto;
-        position: absolute;
         width: 100%;
         --ni-private-listbox-padding: ${smallPadding};
-        max-height: calc(
-            var(--ni-private-select-max-height) - 2 *
-                var(--ni-private-listbox-padding)
-        );
+        max-height: calc(var(--ni-private-select-max-height) - ${smallPadding});
         z-index: 1;
         box-shadow: ${elevation2BoxShadow};
         border: 1px solid ${popupBorderColor};
@@ -155,10 +155,6 @@ export const styles = css`
         padding: var(--ni-private-listbox-padding);
     }
 
-    .listbox[hidden] {
-        display: none;
-    }
-
     :host([open][position='above']) .listbox {
         border-bottom-left-radius: 0;
         border-bottom-right-radius: 0;
@@ -169,12 +165,12 @@ export const styles = css`
         border-top-right-radius: 0;
     }
 
-    :host([open][position='above']) .listbox {
-        bottom: ${controlHeight};
+    :host([open][position='above']) .anchored-region {
+        padding-bottom: ${smallPadding};
     }
 
-    :host([open][position='below']) .listbox {
-        top: calc(${controlHeight} + ${smallPadding});
+    :host([open][position='below']) .anchored-region {
+        padding-top: ${smallPadding};
     }
 
     .selected-value {
@@ -209,16 +205,6 @@ export const styles = css`
 
     :host([disabled]) .indicator slot[name='indicator'] svg {
         fill: ${bodyDisabledFontColor};
-    }
-
-    slot[name='listbox'] {
-        display: none;
-        width: 100%;
-    }
-
-    :host([open]) slot[name='listbox'] {
-        display: flex;
-        position: absolute;
     }
 
     .end {