+++ packages/nimble-components/src/select/index.ts
@@ -1,9 +1,9 @@
-import { attr, html } from '@microsoft/fast-element';
+import { attr, html, observable } from '@microsoft/fast-element';
 import {
+    AnchoredRegion,
     DesignSystem,
     Select as FoundationSelect,
-    SelectOptions,
-    selectTemplate as template
+    SelectOptions
 } from '@microsoft/fast-foundation';
 import { arrowExpanderDown16X16 } from '@ni/nimble-tokens/dist/icons/js';
 import { styles } from './styles';
@@ -11,6 +11,7 @@ import { DropdownAppearance } from '../patterns/dropdown/types';
 import { errorTextTemplate } from '../patterns/error/template';
 import type { ErrorPattern } from '../patterns/error/types';
 import { iconExclamationMarkTag } from '../icons/exclamation-mark';
+import { template } from './template';
 
 declare global {
     interface HTMLElementTagNameMap {
@@ -38,6 +39,12 @@ export class Select extends FoundationSelect implements ErrorPattern {
     @attr({ attribute: 'error-visible', mode: 'boolean' })
     public errorVisible = false;
 
+    /**
+     * @internal
+     */
+    @observable
+    public region?: AnchoredRegion;
+
     // Workaround for https://github.com/microsoft/fast/issues/5123
     public override setPositioning(): void {
         if (!this.$fastController.isConnected) {
@@ -58,6 +65,24 @@ export class Select extends FoundationSelect implements ErrorPattern {
         super.slottedOptionsChanged(prev, next);
         if (value) {
             this.value = value;
+        }
+    }
+
+    private regionChanged(
+        _prev: AnchoredRegion | undefined,
+        _next: AnchoredRegion | undefined
+    ): void {
+        if (this.region && this.control) {
+            this.region.anchorElement = this.control;
+        }
+    }
+
+    private controlChanged(
+        _prev: HTMLElement | undefined,
+        _next: HTMLElement | undefined
+    ): void {
+        if (this.region && this.control) {
+            this.region.anchorElement = this.control;
         }
     }